Move dateformatter to celutil
